+++ isodata/miso.py
@@ -47,3 +47,12 @@ class MISO(ISOBase):
             "time": pd.to_datetime(r[1]['d']).tz_localize(self.default_timezone),
             "demand": float(r[1]['v'].replace(",", ""))
         }
+
+    def get_latest_supply(self):
+        """Returns most recent data point for supply in MW"""
+        mix = self.get_latest_fuel_mix()
+
+        return {
+            "time": mix.time,
+            "supply": mix.total_production
+        }